requirements-test.txt
# ============================================================================
# TESTING DEPENDENCIES
# ============================================================================

# Core testing framework
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Additional testing utilities
factory-boy==3.3.0          # Test data factories
faker==20.1.0               # Generate fake data
freezegun==1.2.2            # Mock datetime
responses==0.24.1           # Mock HTTP requests
vcr.py==5.1.0               # Record HTTP interactions

# Code quality
black==23.11.0              # Code formatter
flake8==6.1.0               # Linting
isort==5.12.0               # Import sorting
mypy==1.7.1                 # Type checking
bandit==1.7.5               # Security linting

# Performance testing  
locust==2.17.0              # Load testing
memory-profiler==0.61.0     # Memory profiling

# Documentation testing
sphinx==7.2.6              # Documentation
sphinx-rtd-theme==1.3.0    # Theme